requirements.txt
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
numpy==1.17.2
pkg-resources==0.0.0
pytz==2019.2
sqlparse==0.3.0
